backend/requirements.txt
langchain==0.3.27
langchain-core==0.3.76
langchain-ollama>=0.2.2,<1.0.0
langchain-text-splitters==0.3.11
tenacity==9.1.2
pydantic==2.11.7
python-dotenv==1.1.1
PyYAML==6.0.2

PyMuPDF==1.26.4
pytesseract==0.3.13
Pillow==11.3.0
pdf2image==1.17.0

chromadb==0.5.23
tiktoken==0.11.0

numpy==1.26.4
pandas==2.2.3

streamlit==1.49.1
tqdm==4.67.1
requests==2.32.3
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
psutil==5.9.8
